Development (#1)
* build (config): Add standard style guide to ESLint configuration

This commits adds the standard style guide to the ESLint configuration by extending the "standard" preset in addition to the existing "next/core-web-vitals" preset.

* build (config): Ignore certificate files and add dynamic sitemap files

* feat (config): Add Storyblok configuration

- Add Storyblok preview token for environment variable
- Allow images from 'a.storyblok.com' domain to be used

* feat (seo): Add next-sitemap configuration file

This commit adds a new file "next-sitemap.config.js" which contains the configuration settings for next-sitemap. It defines the site URL, enables the generation of robots.txt file, and excludes the "/config" path.

* build (config): add necessary npm scripts and packages

* feat (ui): Add favicon

* feat(theme): Add color palette, typography, and theme

- This commit adds a new file `colors.js` to the `theme` directory, which defines a color palette.
- It also adds two new files, `typography.js` and `theme.js`, to the `theme` directory that define the typography and overall theme respectively.
- The `theme.js` file creates a theme object using the `createTheme` function from the `@mui/material/styles` library.

* feat (config): Add page component for dynamic routing

- Created a new component '[...slug].js' to render a page with a title and a StoryblokComponent.
- Added metadata for SEO and social media sharing.
- Implemented the functionality to fetch data from the Storyblok API and return it as props for a Next.js page.
- Designed and implemented the structure of an HTML document in '_document.js'.

* feat(api): add test and subscribe routes

Add a test route to the API that returns a message indicating that the API is working. Also, add a subscribe route that handles a POST request to create a new contact in Twilio Sendgrid. The createNewContact function makes a PUT request to the SendGrid API to add the contact. The routes return appropriate response messages based on the success or failure of the operations.

The test route is added in `src/pages/api/hello.js` and the subscribe route is added in `src/pages/api/subscribe.js`.

* feat (ui): Add ui components  for homepage, blog, poetry, and about pages

This commit adds the necessary UI components for the homepage, blog, poetry, and about page,. Each component is designed to enhance the user experience and provide a visually appealing layout. This commit also includes necessary styling and responsiveness adjustments to ensure optimal display across different devices.

* Refactor (seo); refactor meta description tag on page components

- Set meta description dynamically based on story content from backend in page components.